# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic SDLC Platform API
  description: REST API for the multi-agent SDLC orchestration platform.
  version: 0.4.0
...
paths:
...
  /api/v1/policies:
    get:
      tags:
        - policies
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: Success
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                description: Forbidden
  /api/v1/policies/evaluate:
    post:
      tags:
        - policies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - context_type
                - scope
              properties:
                context_type:
                  type: string
                  enum:
                    - gate
                    - pr
                    - deploy
                    - artifact
                    - schedule
                scope:
                  type: string
                  enum:
                    - global
                    - org
                    - team
                    - repo
                    - sprint
                checks:
                  type: object
              additionalProperties: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: Success
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                description: Validation error
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                description: Forbidden
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                description: Rate limited
  /api/v1/policies/exceptions:
    post:
      tags:
        - policies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - policy_id
                - reason
                - approved_by
              properties:
                policy_id:
                  type: string
                  minLength: 1
                reason:
                  type: string
                  minLength: 1
                approved_by:
                  type: string
                  minLength: 1
                expires:
                  type: string
                scope_override:
                  type: string
              additionalProperties: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: Success
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                description: Validation error
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                description: Forbidden
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                description: Rate limited
...
servers:
  - url: http://localhost:3000
    description: Local development server
...